Title screen finished
Finished the title screen, as simple as it is. Start button now promts you to write your name, and when you do, it takes you to the game. All buttons in the game scene work, and they make you either restart with the same name or get back to the main menu.